Code for DBSCAN, HDBSCAN, and K-means
Dear Dr.Issam,

kindly find attached two notebooks, the first one contains the following:

Explanatory data Analysis
Implementation of DBSCAN
Implementation of HDBSCAN
The second one contains:

Implementation of K-means.
For all algorthims we are using a sample size of 10% and we are limiting the "pm" value to be greater than 10.
We are currently working on implementing the OPTICS algorthim, the geoSilhouette Coefficient, and the writing for the paper.

Kindly have a look at our results and let us know what you think. Your input is highly appreciated.